requirements.txt
astroid==2.2.5
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.2
django-crispy-forms==1.7.2
django-heroku==0.3.1
djangorestframework==3.9.4
gunicorn==19.9.0
isort==4.3.19
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.0.0
psycopg2==2.8.2
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
whitenoise==4.1.2
wrapt==1.11.1